Modify tower side code to use common serialization library
